refactor: move link regex from `UrlFormatter` to `regex.js`
